+++ DonjinKrawler-Client/src/main/java/donjinkrawler/clientpacketcontrol/handlers/ChangeEnemyStrategyPacketHandler.java
@@ -1,6 +1,6 @@
-package donjinkrawler.packetcontrol.handlers;
+package donjinkrawler.clientpacketcontrol.handlers;
 
-import donjinkrawler.packetcontrol.Request;
+import donjinkrawler.clientpacketcontrol.Request;
 import krawlercommon.packets.ChangeEnemyStrategyPacket;
 
 public class ChangeEnemyStrategyPacketHandler extends PacketHandler {
