+++ DonjinKrawler-Client/src/main/java/donjinkrawler/clientpacketcontrol/handlers/CharacterAttackPacketHandler.java
@@ -1,6 +1,6 @@
-package donjinkrawler.packetcontrol.handlers;
+package donjinkrawler.clientpacketcontrol.handlers;
 
-import donjinkrawler.packetcontrol.Request;
+import donjinkrawler.clientpacketcontrol.Request;
 import krawlercommon.packets.CharacterAttackPacket;
 
 public class CharacterAttackPacketHandler extends PacketHandler {
