+++ DonjinKrawler-Client/src/main/java/donjinkrawler/clientpacketcontrol/handlers/MoveCharacterPacketHandler.java
@@ -1,6 +1,6 @@
-package donjinkrawler.packetcontrol.handlers;
+package donjinkrawler.clientpacketcontrol.handlers;
 
-import donjinkrawler.packetcontrol.Request;
+import donjinkrawler.clientpacketcontrol.Request;
 import krawlercommon.packets.MoveCharacter;
 
 public class MoveCharacterPacketHandler extends PacketHandler {
