+++ DonjinKrawler-Client/src/main/java/donjinkrawler/clientpacketcontrol/handlers/RoomPacketHandler.java
@@ -1,6 +1,6 @@
-package donjinkrawler.packetcontrol.handlers;
+package donjinkrawler.clientpacketcontrol.handlers;
 
-import donjinkrawler.packetcontrol.Request;
+import donjinkrawler.clientpacketcontrol.Request;
 import krawlercommon.packets.RoomPacket;
 
 public class RoomPacketHandler extends PacketHandler {
