+++ DonjinKrawler-Server/src/main/java/donjinkrawler/serverpacketcontrol/Request.java
@@ -1,4 +1,4 @@
-package donjinkrawler.pckcontrol;
+package donjinkrawler.serverpacketcontrol;
 
 import com.esotericsoftware.kryonet.Connection;
 import donjinkrawler.GameServer;
