+++ DonjinKrawler-Server/src/main/java/donjinkrawler/serverpacketcontrol/handlers/LoginPacketHandler.java
@@ -1,6 +1,6 @@
-package donjinkrawler.pckcontrol.handlers;
+package donjinkrawler.serverpacketcontrol.handlers;
 
-import donjinkrawler.pckcontrol.Request;
+import donjinkrawler.serverpacketcontrol.Request;
 import krawlercommon.packets.LoginPacket;
 
 public class LoginPacketHandler extends PacketHandler {
